requirements.txt
google-adk==1.27.1
fastapi==0.135.1
uvicorn[standard]==0.41.0
pandas==3.0.1
pydantic==2.12.5
python-dotenv==1.2.2
numpy==2.4.3
vitaldb==1.6.0
